backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.30.1
sqlalchemy==2.0.45
psycopg[binary]>=3.2.2
pydantic==2.9.2
pydantic-settings==2.7.1
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
